[config] separate safety_rules_config
Moving logic around...
